refactor: import affelnet file
