+++ server/src/common/model/schema/afFormation.js
@@ -118,12 +118,6 @@ const afFormationSchema = {
     default: null,
     description: "libellé BAN de la formation",
   },
-  code_cfd: {
-    index: true,
-    type: String,
-    default: null,
-    description: "code formation diplôme de la formation",
-  },
   code_mef: {
     type: String,
     default: null,
